+++ common/users/fused.js
@@ -1,7 +1,6 @@
 const { User: LegacyUser } = process.env.NODE_ENV === 'test' ? require('../../core/_models') : require('../../noncore/_models')
 const { User, Sequelize } = require('../../core/_models')
 const { EmptyResultError } = require('../../common/error-handling/errors')
-const ensureUserSyncedInNeo4j = process.env.NEO4J_ENABLED === 'true' ? require('./legacy/neo4j') : undefined
 
 const userBaseInclude = [{
   include: [{ all: true }]
@@ -49,9 +48,6 @@ async function getIdByGuid (guid) {
 
 async function ensureUserSynced (user) {
   await ensureUserSyncedInTimescaleDB(user)
-  if (ensureUserSyncedInNeo4j) {
-    await ensureUserSyncedInNeo4j(user)
-  }
 }
 
 async function ensureUserSyncedFromToken (req) {